move some files to seperate folders
