+++ lib/enums/image_dimension_enum.dart
@@ -1,0 +1,17 @@
+/// Holds the values from the spotify api for supported Image dimensions
+enum ImageDimension {
+  /// large image
+  large,
+
+  ///medium image
+  medium,
+
+  /// small image
+  small,
+
+  ///xsmall image
+  xSmall,
+
+  /// thumbnail image
+  thumbnail
+}
