+++ lib/enums/repeat_mode_enum.dart
@@ -1,0 +1,11 @@
+/// Holds the values from the spotify api for RepeatModes
+enum RepeatMode {
+  /// repeat is off
+  off,
+
+  /// repeats the current track
+  track,
+
+  /// repeats the current context
+  context,
+}
